course project/requirements.txt
boto3==1.17.78
botocore==1.20.78
Django==2.2.23
django-environ==0.4.5
django-storages==1.11.1
djangorestframework==3.12.4
jmespath==0.10.0
Pillow==8.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-magic==0.4.22
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
